--- src/components/BookingDatePage/BookingDates/BookingDates.module.css
@@ -1,7 +1,0 @@
-.container{
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr;
-    color: #d0d1d0;
-    margin-right: 10px;
-
-}
