--- src/components/Popup/Popup.module.css
@@ -1,29 +1,0 @@
-.popup{
-    position: fixed;
-    top:3em;
-    left:15em;
-    width: 50%;
-    height: 600px;
-    border: 2px solid dodgerblue;
-    border-radius: 10px;
-    background: slategrey;
-
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-
-}
-.popup-inner{
-    position: relative;
-    padding: 32px;
-    width: 100%;
-    max-width: 640px;
-}
-img{
-    width: 150px;
-}
-p{
-    width: 300px;
-}
