+++ src/components/index.js
@@ -1,0 +1,5 @@
+    export * from './Popup';
+    export * from './NaviBar';
+    export * from './MainLayout';
+    export * from './BookingDates'
+    export * from './BookingDate'
