+++ src/index.js
@@ -1,28 +1,23 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import App from './App';
-import  {createGlobalStyle} from "styled-components";
-import {BrowserRouter} from "react-router-dom";
-//import reportWebVitals from './reportWebVitals';
+    import React from 'react';
+    import ReactDOM from 'react-dom/client';
+    import {BrowserRouter} from "react-router-dom";
 
-const Global = createGlobalStyle`
-    *{
-        margin:0;
-        padding:0;
-        box-sizing:border-box;
-    }
-`
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-    <BrowserRouter>
-    <React.StrictMode>
-        <Global/>
-        <App/>
-    </React.StrictMode>
-    </BrowserRouter>
-);
+    import App from './App';
+    import {Global} from "./styled";
+    //import reportWebVitals from './reportWebVitals';
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-//reportWebVitals();
+
+    const root = ReactDOM.createRoot(document.getElementById('root'));
+    root.render(
+        <BrowserRouter>
+        <React.StrictMode>
+            <Global/>
+            <App/>
+        </React.StrictMode>
+        </BrowserRouter>
+    );
+
+    // If you want to start measuring performance in your app, pass a function
+    // to log results (for example: reportWebVitals(console.log))
+    // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+    //reportWebVitals();
