+++ src/services/axiosService.js
@@ -1,9 +1,9 @@
-import axios from "axios";
-import baseURL from "../constants/urls";
+    import axios from "axios";
+    import baseURL from "../constants/urls";
 
 
-const axiosService = axios.create({baseURL});
+    const axiosService = axios.create({baseURL});
 
-export {
-    axiosService
-}
+    export {
+        axiosService
+    }
