+++ src/styled/App.style.js
@@ -1,0 +1,12 @@
+    import styled from "styled-components";
+
+    const AppWrapper = styled.div`
+    width:100%;
+    min-height: 150vh;
+    background: rgb(62,58,180);
+    background: linear-gradient(90deg, rgba(62,58,180,1) 0%, rgba(253,29,29,1) 80%, rgba(252,176,69,1) 100%);
+    `
+
+    export {
+        AppWrapper
+    }
