+++ src/styled/Global.style.js
@@ -1,0 +1,13 @@
+    import {createGlobalStyle} from "styled-components";
+
+    const Global = createGlobalStyle`
+            *{
+            margin:0;
+            padding:0;
+            box-sizing:border-box;
+        }
+    `
+
+    export {
+        Global
+    }
