+++ src/styled/index.js
@@ -1,0 +1,4 @@
+    export * from './App.style';
+    export * from './BookingDates.style';
+    export * from './Popup.style';
+    export * from './Global.style';
